Break out rest integration tests into full integration and dbquery files. Update spec files to specify their setup steps
