+++ hedera-mirror-rest/__tests__/specs/accounts-01-no-args.spec.json
@@ -1,14 +1,58 @@
 {
+  "description": "Account api calls for all accounts",
+  "setup": {
+    "accounts": [
+      {
+        "entity_num": 1,
+        "exp_time_ns": null
+      },
+      {
+        "entity_num": 2,
+        "exp_time_ns": null
+      },
+      {
+        "entity_num": 4,
+        "exp_time_ns": null
+      },
+      {
+        "entity_num": 5,
+        "exp_time_ns": null
+      },
+      {
+        "entity_num": 6,
+        "exp_time_ns": null
+      },
+      {
+        "entity_num": 7,
+        "exp_time_ns": null
+      },
+      {
+        "entity_num": 8,
+        "exp_time_ns": null
+      },
+      {
+        "entity_num": 9,
+        "exp_time_ns": "-9223372036854775808"
+      },
+      {
+        "entity_num": 10,
+        "exp_time_ns": "9223372036854775807"
+      }
+    ],
+    "balances": [],
+    "transactions": [],
+    "cryptotransfers": []
+  },
   "url": "/api/v1/accounts",
   "responseStatus": 200,
   "responseJson": {
     "accounts": [
       {
         "balance": {
-          "timestamp": "0.000002000",
+          "timestamp": "0.000000001",
           "balance": 10
         },
-        "account": "0.15.1",
+        "account": "0.0.1",
         "expiry_timestamp": null,
         "auto_renew_period": null,
         "key": null,
@@ -16,10 +60,10 @@
       },
       {
         "balance": {
-          "timestamp": "0.000002000",
+          "timestamp": "0.000000002",
           "balance": 20
         },
-        "account": "0.15.2",
+        "account": "0.0.2",
         "expiry_timestamp": null,
         "auto_renew_period": null,
         "key": null,
@@ -27,10 +71,10 @@
       },
       {
         "balance": {
-          "timestamp": "0.000002000",
+          "timestamp": "0.000000003",
           "balance": 30
         },
-        "account": "0.15.3",
+        "account": "0.0.3",
         "expiry_timestamp": null,
         "auto_renew_period": null,
         "key": null,
@@ -38,10 +82,10 @@
       },
       {
         "balance": {
-          "timestamp": "0.000002000",
+          "timestamp": "0.000000004",
           "balance": 40
         },
-        "account": "0.15.4",
+        "account": "0.0.4",
         "expiry_timestamp": null,
         "auto_renew_period": null,
         "key": null,
@@ -49,10 +93,10 @@
       },
       {
         "balance": {
-          "timestamp": "0.000002000",
+          "timestamp": "0.000000005",
           "balance": 50
         },
-        "account": "0.15.5",
+        "account": "0.0.5",
         "expiry_timestamp": null,
         "auto_renew_period": null,
         "key": null,
@@ -60,10 +104,10 @@
       },
       {
         "balance": {
-          "timestamp": "0.000002000",
+          "timestamp": "0.000000006",
           "balance": 60
         },
-        "account": "0.15.6",
+        "account": "0.0.6",
         "expiry_timestamp": null,
         "auto_renew_period": null,
         "key": null,
@@ -71,10 +115,10 @@
       },
       {
         "balance": {
-          "timestamp": "0.000002000",
+          "timestamp": "0.000000007",
           "balance": 70
         },
-        "account": "0.15.7",
+        "account": "0.0.7",
         "expiry_timestamp": null,
         "auto_renew_period": null,
         "key": null,
@@ -82,10 +126,10 @@
       },
       {
         "balance": {
-          "timestamp": "0.000002000",
+          "timestamp": "0.000000008",
           "balance": 80
         },
-        "account": "0.15.8",
+        "account": "0.0.8",
         "expiry_timestamp": null,
         "auto_renew_period": null,
         "key": null,
@@ -93,10 +137,10 @@
       },
       {
         "balance": {
-          "timestamp": "0.000002000",
+          "timestamp": "0.000000009",
           "balance": 90
         },
-        "account": "0.15.9",
+        "account": "0.0.9",
         "expiry_timestamp": "-9223372036.854775808",
         "auto_renew_period": null,
         "key": null,
@@ -104,10 +148,10 @@
       },
       {
         "balance": {
-          "timestamp": "0.000002000",
+          "timestamp": "0.000000010",
           "balance": 100
         },
-        "account": "0.15.10",
+        "account": "0.0.10",
         "expiry_timestamp": "9223372036.854775807",
         "auto_renew_period": null,
         "key": null,
